+++ tests_integration/test_data/projects/napp_app_manifest_v2/app/setup_script.sql
@@ -1,0 +1,3 @@
+-- This is the v2 version of the napp_init_v1 project
+
+CREATE OR ALTER VERSIONED SCHEMA core;